move main.dart to Pages
