+++ test_driver/app.dart
@@ -1,4 +1,4 @@
-import 'package:finesse_nation/main.dart' as app;
+import 'package:finesse_nation/Pages/main.dart' as app;
 import 'package:flutter_driver/driver_extension.dart';
 
 void main() {
